Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,25 +1,25 @@
 #include <stdio.h>
 int main(){
     int a,b;
-    float d;
+    int d;
     char c;
     scanf("%d %d %c",&a,&b,&c);
     switch(c){
         case '+':
         d=a+b;
-        printf("%f",d);
+        printf("%d",d);
         break;
         case '-':
         d=a-b;
-        printf("%f",d);
+        printf("%d",d);
         break;
         case '*':
         d=a*b;
-        printf("%f",d);
+        printf("%d",d);
         break;
         case '/':
         d=a/b;
-        printf("%f",d);
+        printf("%d",d);
         break;
         default:
         printf("error");
